kafka/requirements.txt
kafka-python==2.0.2
python-dotenv==0.19.2
requests
PyYAML==6.0.2